+++ plug-ins/Apple.CoreHaptics/Apple.CoreHaptics_Unity/Assets/Apple.CoreHaptics/CHANGELOG.md
@@ -38,7 +38,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.0.1] - 2022-10-10
 ### Changed
-- Updated package name to `com.dirtybit.apple.unityplugin.corehaptics`
+- Updated package name to `com.apple.unityplugin.corehaptics`
 
 ## [1.0.0] - 2022-06-02
 ### Added