+++ plug-ins/Apple.GameController/Apple.GameController_Unity/Assets/Apple.GameController/CHANGELOG.md
@@ -33,7 +33,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.0.1] - 2022-10-10
 ### Changed
-- Updated package name to `com.dirtybit.apple.unityplugin.gamecontroller`
+- Updated package name to `com.apple.unityplugin.gamecontroller`
 
 ## [1.0.0] - 2022-06-02
 ### Added